+++ src/rest/middlewares/index.ts
@@ -11,4 +11,5 @@ export * from './headerParameter.middleware';
 export * from './headers.middleware';
 export * from './cache.middleware';
 export * from './logger.middleware';
+export * from './mockLogger.middleware';
 export * from './progressIndicatorGroup.middleware';